Refactor config and exploit handling, centralize version constant
- Remove exploit name from config manager and related methods - Move
exploit output types and status constants to exploit/types.go - Use
models.VERSION for CLI and server versioning - Remove debug flag from
TUI startup and main - Minor comment and style improvements
