+++ cmd/client/cmd/exploit.go
@@ -298,8 +298,6 @@ func test(cmd *cobra.Command, args []string) {
 
 // create handles the creation of a new exploit template
 func create(cmd *cobra.Command, args []string) {
-	cm := config.GetConfigManager()
-	cm.SetExploitName(exploitName)
 	res, err := exploit.Create(exploitName)
 	if err != nil {
 		logger.Log.Error().Err(err).Msg("Error creating exploit")
@@ -327,8 +325,6 @@ func list(cmd *cobra.Command, args []string) {
 
 // remove handles the removal of an exploit template
 func remove(cmd *cobra.Command, args []string) {
-	cm := config.GetConfigManager()
-	cm.SetExploitName(exploitName)
 	res, err := exploit.Remove(exploitName)
 	if err != nil {
 		logger.Log.Error().Err(err).Msg("Error removing exploit")
